Reorganize event admin by category
