+++ app/javascript/EventAdmin/EventAdminRunsTable.jsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { Route, Link } from 'react-router-dom';
 import { pluralize } from 'inflected';
@@ -7,32 +7,16 @@ import EditRun from './EditRun';
 import EventAdminRow from './EventAdminRow';
 import { EventAdminEventsQuery } from './queries.gql';
 import useQuerySuspended from '../useQuerySuspended';
-import { sortByLocaleString } from '../ValueUtils';
 import ErrorDisplay from '../ErrorDisplay';
 import usePageTitle from '../usePageTitle';
 import useValueUnless from '../useValueUnless';
 import buildEventCategoryUrl from './buildEventCategoryUrl';
-
-const getNormalizedEventTitle = event => event.title
-  .replace(/^(the|a|) /i, '')
-  .replace(/[^A-Za-z0-9]/g, '')
-  .toLocaleLowerCase();
+import useEventAdminCategory from './useEventAdminCategory';
 
 function EventAdminRunsTable({ eventCategoryId }) {
   const { data, error } = useQuerySuspended(EventAdminEventsQuery);
 
-  const filteredEvents = useMemo(
-    () => (error ? [] : data.events.filter(event => event.event_category.id === eventCategoryId)),
-    [data, error, eventCategoryId],
-  );
-  const sortedEvents = useMemo(
-    () => sortByLocaleString(filteredEvents, getNormalizedEventTitle),
-    [filteredEvents],
-  );
-  const eventCategory = useMemo(
-    () => (error ? null : data.convention.event_categories.find(c => c.id === eventCategoryId)),
-    [data, error, eventCategoryId],
-  );
+  const [eventCategory, sortedEvents] = useEventAdminCategory(data, error, eventCategoryId);
 
   usePageTitle(useValueUnless(() => pluralize(eventCategory.name), error));
 
@@ -43,7 +27,8 @@ function EventAdminRunsTable({ eventCategoryId }) {
   return (
     <div>
       <Link to={`${buildEventCategoryUrl(eventCategory)}/new`} className="btn btn-primary mt-4 mb-2">
-        New event
+        {'Create new '}
+        {eventCategory.name.toLowerCase()}
       </Link>
 
       <table className="table table-striped no-top-border">
