+++ app/javascript/EventAdmin/RecurringEventAdmin.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Link, Route } from 'react-router-dom';
+import { pluralize } from 'inflected';
 
 import EditRun from './EditRun';
 import RecurringEventSection from './RecurringEventSection';
@@ -7,38 +9,35 @@ import { EventAdminEventsQuery } from './queries.gql';
 import useQuerySuspended from '../useQuerySuspended';
 import ErrorDisplay from '../ErrorDisplay';
 import usePageTitle from '../usePageTitle';
+import useEventAdminCategory from './useEventAdminCategory';
+import buildEventCategoryUrl from './buildEventCategoryUrl';
+import useValueUnless from '../useValueUnless';
 
-function RecurringEventAdmin() {
+function RecurringEventAdmin({ eventCategoryId }) {
   const { data, error } = useQuerySuspended(EventAdminEventsQuery);
+  const [eventCategory, sortedEvents] = useEventAdminCategory(data, error, eventCategoryId);
 
-  usePageTitle('Recurring Events');
+  usePageTitle(useValueUnless(() => pluralize(eventCategory.name), error));
 
   if (error) {
     return <ErrorDisplay graphQLError={error} />;
   }
 
-  const volunteerEvents = data.events.filter((event) => {
-    const eventCategory = data.convention.event_categories
-      .find(c => c.id === event.event_category.id);
-    return eventCategory.scheduling_ui === 'recurring' && event.status === 'active';
-  });
-  volunteerEvents.sort((a, b) => a.title.localeCompare(b.title, { sensitivity: 'base' }));
-  const eventSections = volunteerEvents.map(event => (
-    <RecurringEventSection
-      convention={data.convention}
-      event={event}
-      key={event.id}
-    />
-  ));
-
   return (
     <div>
-      <Link className="btn btn-primary mt-4" to="/admin_events/recurring_events/new">
-        Create new recurring event
+      <Link className="btn btn-primary mt-4" to={`${buildEventCategoryUrl(eventCategory)}/new`}>
+        {'Create new '}
+        {eventCategory.name.toLowerCase()}
       </Link>
       <hr className="my-4" />
-      {eventSections}
-      <Route path="/admin_events/recurring_events/:eventId/runs/:runId/edit">
+      {sortedEvents.map(event => (
+        <RecurringEventSection
+          convention={data.convention}
+          event={event}
+          key={event.id}
+        />
+      ))}
+      <Route path={`${buildEventCategoryUrl(eventCategory)}/:eventId/runs/:runId/edit`}>
         {props => (
           <EditRun {...props} events={data.events} convention={data.convention} />
         )}
@@ -46,5 +45,9 @@ function RecurringEventAdmin() {
     </div>
   );
 }
+
+RecurringEventAdmin.propTypes = {
+  eventCategoryId: PropTypes.number.isRequired,
+};
 
 export default RecurringEventAdmin;
