--- app/javascript/EventAdmin/SingleRunEventAdmin.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-
-import NewSingleRunEvent from './NewSingleRunEvent';
-import EditSingleRunEvent from './EditSingleRunEvent';
-import SingleRunEventAdminList from './SingleRunEventAdminList';
-
-function SingleRunEventAdmin() {
-  return (
-    <Switch>
-      <Route path="/admin_events/filler_events/new" component={NewSingleRunEvent} />
-      <Route path="/admin_events/filler_events/:id/edit" component={EditSingleRunEvent} />
-      <Route component={SingleRunEventAdminList} />
-    </Switch>
-  );
-}
-
-export default SingleRunEventAdmin;
